// ts/svelte/free2z/src/lib/api/f2z.schemas.ts
export type AccessEnum = typeof AccessEnum[keyof typeof AccessEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccessEnum = {
  private: 'private',
  members: 'members',
  public: 'public',
} as const;

export type BlankEnum = typeof BlankEnum[keyof typeof BlankEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BlankEnum = {
  '': '',
} as const;
// ...
export type CategoryEnum = typeof CategoryEnum[keyof typeof CategoryEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CategoryEnum = {
  ART: 'ART',
  COMEDY: 'COMEDY',
  COMMUNITY: 'COMMUNITY',
  CHARITY: 'CHARITY',
  CRYPTO: 'CRYPTO',
  EDUCATION: 'EDUCATION',
  FICTION: 'FICTION',
  FOR_TRADE: 'FOR TRADE',
  FREE2Z: 'FREE2Z',
  FUNDRAISING: 'FUNDRAISING',
  GAMING: 'GAMING',
  HEALTH: 'HEALTH',
  LIFESTYLE: 'LIFESTYLE',
  MATH: 'MATH',
  MUSIC: 'MUSIC',
  PODCAST: 'PODCAST',
  POLITICS: 'POLITICS',
  RELIEF: 'RELIEF',
  SCIENCE: 'SCIENCE',
  SERVICE: 'SERVICE',
  SPORTS: 'SPORTS',
  TECHNOLOGY: 'TECHNOLOGY',
  ZCASH: 'ZCASH',
} as const;
// ...
export interface FeaturedImage {
  readonly id: number;
  readonly url: string;
  readonly name: string;
  readonly thumbnail: string;
  readonly created_at: string;
  readonly updated_at: string;
  /** @maxLength 128 */
  title?: string;
  /** @maxLength 255 */
  description?: string;
  access?: AccessEnum;
  mime_type: MimeTypeEnum;
}
// ...
export type MimeTypeEnum = typeof MimeTypeEnum[keyof typeof MimeTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MimeTypeEnum = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/gif': 'image/gif',
  'image/svg+xml': 'image/svg+xml',
  'image/webp': 'image/webp',
  'video/mp4': 'video/mp4',
} as const;

export interface Page {
  /** @maxLength 512 */
  prompt: string;
  /** @maxLength 3072 */
  content?: string;
  readonly image_url: string;
  /**
   * @minimum 0
   * @maximum 2147483647
   */
  order: number;
  readonly created_at: string;
  readonly updated_at: string;
}

export interface PageComment {
  readonly created_at: string;
  /** @maxLength 256 */
  comment: string;
}
// ...
export interface SimpleCreator {
  /**
   * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
   * @maxLength 150
   * @pattern ^[\w.@+-]+$
   */
  username: string;
  avatar_image: FeaturedImage;
  /** @maxLength 255 */
  p2paddr?: string;
  /**
   * Price for 30 days of membership, livestream access, etc
   * @nullable
   * @pattern ^-?\d{0,6}(?:\.\d{0,0})?$
   */
  member_price?: string | null;
}
// ...
export const ZPageDetailCategory = {...CategoryEnum,...BlankEnum,} as const
/**
 * Only for retrieval
 */
export interface ZPageDetail {
  /** UUID to get the zPage */
  readonly free2zaddr: string;
  /**
   * Optional vanity URL field
   * @maxLength 128
   * @nullable
   * @pattern ^[-a-zA-Z0-9_]+$
   */
  vanity?: string | null;
  p2paddr: string;
  /** @maxLength 128 */
  title: string;
  /** @maxLength 100000 */
  content: string;
  category?: typeof ZPageDetailCategory[keyof typeof ZPageDetailCategory] ;
  is_published?: boolean;
  is_subscriber_only?: boolean;
  readonly total_raised: number;
  featured_image: FeaturedImage;
  /**
   * Description for meta tag, AI if blank
   * @maxLength 255
   */
  description?: string;
  is_verified?: boolean;
  /** @pattern ^-?\d{0,2}(?:\.\d{0,8})?$ */
  total?: string;
  /**
   * Cache of rank calculation
   * @pattern ^-?\d{0,12}(?:\.\d{0,8})?$
   */
  readonly f2z_score: string;
  readonly comments: readonly PageComment[];
  creator: SimpleCreator;
  readonly created_at: string;
  readonly updated_at: string;
  readonly get_url: string;
  tags: string[];
  /** @nullable */
  publish_at?: string | null;
}